==> step1-load-by-Seurat-v5.R <==
rm(list=ls())
source('scRNA_scripts/lib.R')
getwd()

#### txt/csv/tsv data 读取###
# 一般下载下来的都是tar结尾的压缩文件
#untar("GSE155698_RAW.tar",exdir = "GSE155698_RAW")
dir='GSE155698_RAW' 
fs=list.files('GSE155698_RAW/','^GSM')
fs
# 使用grep筛选出不包含"PBMC"的文件名
setwd('GSE155698_RAW/')
files_to_remove <- grep("PBMC", fs, value = TRUE)
file.remove(files_to_remove)
setwd('../')
library(tidyverse)
fs=list.files('GSE155698_RAW/','^GSM')
fs
samples=str_split(fs,'_',simplify = T)[,1]

# ##处理数据，将原始文件分别整理为barcodes.tsv.gz，features.tsv.gz和matrix.mtx.gz到各自的文件夹
# #批量将文件名改为 Read10X()函数能够识别的名字
# if(F){
#   lapply(unique(samples),function(x){
#     # x = unique(samples)[1]
#     y=fs[grepl(x,fs)]
#     folder=paste0("GSE155698/GSE155698_RAW/", paste(str_split(y[1],'_',simplify = T)[,1:2], collapse = "_"))
#     dir.create(folder,recursive = T)
#     #为每个样本创建子文件夹
#     file.rename(paste0("GSE155698/GSE155698_RAW/",y[1]),file.path(folder,"barcodes.tsv.gz"))
#     #重命名文件，并移动到相应的子文件夹里
#     file.rename(paste0("GSE155698/GSE155698_RAW/",y[2]),file.path(folder,"features.tsv.gz"))
#     file.rename(paste0("GSE155698/GSE155698_RAW/",y[3]),file.path(folder,"matrix.mtx.gz"))
#   })
# }

setwd('GSE155698_RAW/')
if(length(fs) > 0) {
  lapply(unique(samples), function(x) {
    #x = unique(samples)[1]
    y <- fs[grepl(x, fs)]
    folder <- paste0("GSE155698_data/", paste(str_split(y[1], '_', simplify = TRUE)[, 1:3], collapse = "_"))
    dir.create(folder, recursive = TRUE)
    untar(y[1], exdir = folder)
    subfolder <- list.dirs(folder, full.names = TRUE, recursive = FALSE)[1]
    subsubfolder = list.dirs(subfolder, full.names = TRUE, recursive = FALSE)[1]
    files_in_subfolder <- list.files(subsubfolder, full.names = TRUE)
    
    # 移动三个文件到第一层级
    for (file_path in files_in_subfolder) {
      file.rename(file_path, file.path(folder, basename(file_path)))
    }
    
    # 删除多余文件夹
    unlink(subfolder, recursive = TRUE)
  })
}
###检查数据，可以直接在linux中使用tree来查看层级结构
setwd('../')
dir='GSE155698_RAW/GSE155698_data/' 
samples=list.files( dir )
samples 
sceList = lapply(samples,function(pro){ 
  # pro=samples[1] 
  print(pro)  
  tmp = Read10X(file.path(dir,pro )) 
  if(length(tmp)==2){
    ct = tmp[[1]] 
  }else{ct = tmp}
  sce =CreateSeuratObject(counts =  ct ,
                          project =  pro  ,
                          min.cells = 5,
                          min.features = 300 )
  return(sce)
}) 

do.call(rbind,lapply(sceList, dim)) 

sce.all=merge(x=sceList[[1]],
              y=sceList[ -1 ],   add.cell.ids = samples ) 
names(sce.all@assays$RNA@layers)
sce.all[["RNA"]]$counts[1:5,1:5]
LayerData(sce.all,assay = "RNA", layer = "counts")
sce.all <- JoinLayers(sce.all)
dim(sce.all[["RNA"]]$counts )

as.data.frame(sce.all@assays$RNA$counts[1:10, 1:2])
head(sce.all@meta.data, 10)
table(sce.all$orig.ident) 
sce.all

##orig.ident(样品名),nCount_RNA(每个细胞中所有基因的测序读段总数。衡量每个细胞测序数据总量的指标)
#nFeature_RNA（每个细胞中检测到的基因数量）

sp='human'
# 如果为了控制代码复杂度和行数 
# 可以省略了质量控制环节
###### step2: QC质控 ######
#目的是为了去除质量较差细胞，低质量细胞会形成独特的亚群，使分群结果变得复杂
dir.create("1-QC")
setwd("1-QC")
# 如果过滤的太狠，就需要去修改这个过滤代码
source('../scRNA_scripts/qc.R')
sce.all.filt = basic_qc(sce.all)
print(dim(sce.all))
print(dim(sce.all.filt))
setwd('../')
getwd()

###### step3: harmony整合多个单细胞样品 ######

if(T){
  dir.create("2-harmony")
  getwd()
  setwd("2-harmony")
  source('../scRNA_scripts/harmony.R')
  # 默认 ScaleData 没有添加"nCount_RNA", "nFeature_RNA"
  # 默认的
  sce.all.int = run_harmony(sce.all.filt)
  setwd('../') 
}


###### step4:  看标记基因库 ######
# 原则上分辨率是需要自己肉眼判断，取决于个人经验
# 为了省力，我们直接看  0.1和0.8即可

table(Idents(sce.all.int))
table(sce.all.int$seurat_clusters)
table(sce.all.int$RNA_snn_res.0.1) 
table(sce.all.int$RNA_snn_res.0.3) 
table(sce.all.int$RNA_snn_res.0.5) 

# getwd()
# dir.create('check-by-0.1')
# setwd('check-by-0.1')
# sel.clust = "RNA_snn_res.0.1"
# sce.all.int <- SetIdent(sce.all.int, value = sel.clust)
# table(sce.all.int@active.ident)
# 
# source('../scRNA_scripts/check-all-markers.R')
# setwd('../')
# getwd()

dir.create('check-by-0.5')
setwd('check-by-0.5')
sel.clust = "RNA_snn_res.0.5"
sce.all.int <- SetIdent(sce.all.int, value = sel.clust)
table(sce.all.int@active.ident) 
source('../scRNA_scripts/check-all-markers.R')
setwd('../') 
getwd()

# dir.create('check-by-0.8')
# setwd('check-by-0.8')
# sel.clust = "RNA_snn_res.0.8"
# sce.all.int <- SetIdent(sce.all.int, value = sel.clust)
# table(sce.all.int@active.ident) 
# source('../scRNA_scripts/check-all-markers.R')
load(file = 'qc-_marker_cosg.Rdata')
top_10 <- unique(as.character(apply(marker_cosg$names,2,head,10))) 

library(dplyr)  
sink(file = 'for_act.txt')
cat(paste0('cluster',colnames(marker_cosg$names),':',
           unlist(apply(marker_cosg$names,2,function(x){
             paste(head(x),collapse=',')
           })),'\n'))
sink()
#a=data.table::fread('ACT_Annotation results_top 1.txt')
write.csv(marker_cosg$names,file = 'marker_cosg.csv') 
setwd('../') 
getwd()

last_markers_to_check


###### step5: 确定单细胞亚群生物学名字 ######
# 一般来说，为了节省工作量，我们选择0.1的分辨率进行命名
# 因为命名这个步骤是纯人工 操作
# 除非0.1确实分群太粗狂了，我们就选择0.8 
 
## 如果是视网膜
if(F){
  
  data=read.table("./paper.retina-cell_marker.txt",sep="\t",header=T)
  head(data)
  data$gene <- str_to_title(data$gene)
  selected_genes = split(data$gene,data$cell)
  selected_genes
  selected_genes = lapply(selected_genes, str_to_title)
  selected_genes
  p3 <- DotPlot(object = sce.all.int, features = selected_genes, 
                group.by = 'RNA_snn_res.0.1',
                assay = "RNA") +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=0.5)
    )+theme(
      strip.text= element_text(angle = 90, vjust = 0.5, hjust=0.5)
    )
  p3
  ggsave(p3,filename = "check_retina_markers_by_RNA_snn_res.0.1.pdf",
         units = "cm",width = 36,height = 20) 
  p3 <- DotPlot(object = sce.all.int, features = selected_genes, 
                group.by = 'RNA_snn_res.0.8',
                assay = "RNA") +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=0.5)
    )+theme(
      strip.text= element_text(angle = 90, vjust = 0.5, hjust=0.5)
    )
  p3
  ggsave(p3,filename = "check_retina_markers_by_RNA_snn_res.0.8.pdf",
         units = "cm",width = 36,height = 20) 
} 

## 如果是皮肤
if(F){
  
  # keratinocyte：SFN
  #melanocyte1 "MLANA" 
  selected_genes=c('MUC4', 'PI3', 'SIX3', # nose
                   'SCGB1A1', 'TFF3',   'KRT1',   'KRT10',    
                   'KRT5', 'TP63',   'DLK2',
                   'MKI67', 'TOP2A', 'CDC20' ,
                   'DMD','PUS7','PGAP1','IFI44L',"MLANA" ,'SFN',
                   'MUC5AC' , 'MUC5B',# secretory cells
                   'FOXJ1', 'TPPP3',  'SNTN',  # multiciliated cells 
                   'DEUP1', 'FOXN4',  'CDC20B' # deuterosomal cells
  )
  p1 <- DotPlot(sce.all.int, features = unique(str_to_upper(selected_genes)),
                assay='RNA' ,group.by = 'RNA_snn_res.0.1'  )  + coord_flip() # +th
  
  p1
  
  gplots::balloonplot(table(sce.all.int$RNA_snn_res.0.8,sce.all.int$orig.ident))
  gplots::balloonplot(table(sce.all.int$RNA_snn_res.0.1,sce.all.int$orig.ident))
  
}

## 如果是大脑
if(F){
  
  astrocytes = c("AQP4", "ADGRV1", "GPC5", "RYR3") 
  endothelial = c("CLDN5", "ABCB1", "EBF1") 
  excitatory = c("CAMK2A", "CBLN2", "LDB2") 
  inhibitory = c("GAD1", "LHFPL3", "PCDH15") 
  microglia = c("C3", "LRMDA", "DOCK8") 
  oligodendrocytes = c("MBP", "PLP1", "ST18")
  # 下面的 OPC是 上面的 oligodendrocytes 的前体细胞 
  OPC='Tnr,Igsf21,Neu4,Gpr17'
  Ependymal='Cfap126,Fam183b,Tmem212,pifo,Tekt1,Dnah12'
  pericyte=c(  'DCN', 'LUM',  'GSN' ,'FGF7','MME', 'ACTA2','RGS5')
  
  gene_list = list(
    Astro = astrocytes,
    Endo = endothelial,
    Excit = excitatory,
    Inhib = inhibitory,
    Mic = microglia,
    Oligo = oligodendrocytes,
    OPC= str_to_upper(trimws(strsplit(OPC,',')[[1]])),
    Ependymal= str_to_upper(trimws(strsplit(Ependymal,',')[[1]])) ,
    peri = pericyte
  )
  gene_list = lapply(gene_list , str_to_title)
  
  p2 = DotPlot( sce.all.int,  features = gene_list, 
                group.by = 'RNA_snn_res.0.1') + 
    theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
  p2    
  ggsave('brain-0.1.pdf',width=12)
  p2 = DotPlot( sce.all.int,  features = gene_list, 
                group.by = 'RNA_snn_res.0.8') + 
    theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
  p2    
  ggsave('brain-0.8.pdf',width=12)
  
  gplots::balloonplot(table(sce.all.int$RNA_snn_res.0.8,sce.all.int$orig.ident))
  gplots::balloonplot(table(sce.all.int$RNA_snn_res.0.1,sce.all.int$orig.ident))
  
  Astrocytes = c("Slc1a2", "Slc1a3","Gpc5", "Prex2","Atp1a2") 
  Epithelial = c("Kitl","Thsd4","Gbp3","Rps8","Tshz2") 
  Oligodendrosytes = c("Mbp","St18","Rnf220","Plp1","Slc24a2") 
  Ependymal = c( "Dnah6","Dnah12","Cfap44","Spag17","Spag16") 
  Oligodendrosyte_progenitor = c("Lhfpl3","Tnr","Dscam","Xylt1","Pcdh15") 
  Contaminating_neurons = c("Meg3","Snhg11","Syt1","Cntnap2","Nrg3") 
  Immune_cells = c( "Ctsd","Hexb","Inpp5d","Ccr5","Cx3cr1")
  Choroid_plexus_epithelial_cells = c("Htr2c","Ttr","Otx2os1","Enpp2","Trpm3")
  Meningeal_cells=c("Fbxl7","Eya2","Bnc2","Cped1","Tmtc1")
  Endothelial =c("Flt1","Slco1a4","Mecom","Adgrl4","Cldn5")
  
  gene_list = list(Astrocytes,Epithelial,Oligodendrosytes,Ependymal,Oligodendrosyte_progenitor,
                   Contaminating_neurons,Immune_cells,Choroid_plexus_epithelial_cells,Meningeal_cells,Endothelial)
  
  names(gene_list)=trimws( strsplit('Astrocytes,Epithelial,Oligodendrosytes,Ependymal,Oligodendrosyte_progenitor,
                   Contaminating_neurons,Immune_cells,Choroid_plexus_epithelial_cells,Meningeal_cells,Endothelial',',')[[1]])
  p_all_markers=DotPlot(sce.all.int, 
                        group.by = 'RNA_snn_res.0.1',
                        features = gene_list,
                        scale = T,assay='RNA' )+  
    theme(axis.text.x=element_text(angle=45,hjust = 1))
  p_all_markers
  ggsave('check_paper_markers_RNA_snn_res.0.1.pdf',
         height = 8,width = 12)
  
}

## 如果是肾脏 
if(F){
  # proximal tubule (PT)
  # parietal epithelial cells (PEC)
  # thick ascending limb (TAL)
  # distal tubule (DCT1, DCT2)
  # connecting tubule (CNT)
  # collecting duct (PC, ICA, ICB)
  # endothelial cells (ENDO)
  # glomerular cell types (MES, PODO)
  # fibroblasts (FIB)
  # and a small population of leukocytes (LEUK) 
  
  celltype.markers <- c("CUBN","HAVCR1","SLC5A1","SLC5A2", # PT and PT-VCAM1+ markers
                        "CFH", # PEC
                        "SLC12A1", # TAL NKCC2
                        "SLC12A3","TRPM6", # DCT1 and DCT2 NCC
                        "SCNN1G","TRPV5", # DCT2/CNT ENaC
                        "CALB1", # CNT
                        "AQP2", # PC
                        "ATP6V0D2", # ICA and ICB
                        "SLC4A1","SLC26A7", # ICA
                        "SLC26A4", # ICB
                        "NPHS1","NPHS2", # PODO
                        "PECAM1","FLT1","EMCN", # ENDO
                        "CLDN5", # GEC
                        "ITGA8","PDGFRB", # MES
                        "ACTA2","CALD1", # FIB
                        "PTPRC") # WBC
  p2 <- DotPlot(sce.all.int, features = celltype.markers) + 
    theme(axis.text.x = element_text(angle = 90, hjust = 1))
  
  p1 <- DotPlot(sce.all.int, features = unique(str_to_upper(celltype.markers)),
                assay='RNA' ,group.by = 'RNA_snn_res.0.1'  )  + coord_flip() # +th
  
  p1
  
  p2 = DotPlot( sce.all.int,  features = celltype.markers, 
                group.by = 'RNA_snn_res.0.1') + 
    theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
  p2    
  ggsave('kidney-0.1.pdf',width=12)
  p2 = DotPlot( sce.all.int,  features = celltype.markers, 
                group.by = 'RNA_snn_res.0.8') + 
    theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
  p2    
  ggsave('kidney-0.8.pdf',width=12)
  
  gplots::balloonplot(table(sce.all.int$RNA_snn_res.0.8,sce.all.int$orig.ident))
  gplots::balloonplot(table(sce.all.int$RNA_snn_res.0.1,sce.all.int$orig.ident))
  
}

# 如果是胰腺 
if(F){
  # acinar ROIs (CEL, CPA1, CELA2B, and PRSS3)
  # ductal ROIs (MUC5B and MUC20).
  selected_genes=  c("MUC5AC", "TFF1",
                     'CEL', 'CPA1', 'CELA2B',   'PRSS3',
                     'MUC5B' , 'MUC20',
                     "MUC6", "TFF2",
                     "CHGB", "CHGA", # enteroendocrine cells (marked with CHGA;
                     "PGA3","PGA4",
                     "MKI67","TOP2A","BIRC5",
                     "MUC2","ITLN1","SPINK4",
                     "APOA1","ALPI","FABP1","APOA4",
                     "CEACAM5","CEACAM6",
                     "CD79A","CD19",
                     "CD2", "CD3D","CD3E","CD3G",
                     "DCN","PDPN",
                     "ACTA2","ACTN2","MYL2","MYH2",
                     "VWF","ENG",
                     "TPSAB1",
                     "CXCL3","IL8",
                     "OLFM4","PIGR","EPHB2","SOX9",
                     "CD14","CD163","CD68","CSF1R")
  p1 <- DotPlot(sce.all.int, features = unique(str_to_upper(selected_genes)),
                assay='RNA' ,group.by = 'RNA_snn_res.0.1'  )  + coord_flip() # +th
  
  p1
  ggsave('paper-markers-pancr.pdf',height =  8)
  
  gene_list <- list(
    "β cells" = c("INS", "PCSK1", "G6PC2"),
    "α cells" = c("GCG"),
    "δ cells" = c("SST"),
    "ε cells" = c("GHRL"),
    "pancreatic progenitor cells" = c("SOX9"),
    "proliferation cells" = c("MKI67"),
    "EC cells" = c("FEV")
    #, "polyhormonal endocrine cells" = c("GCG", "INS")
  )
  
  p2 = DotPlot( sce.all.int,  features = gene_list, 
                group.by = 'RNA_snn_res.0.1') + 
    theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
  p2    
  ggsave('pancreatic-0.1.pdf',width=12)
  p2 = DotPlot( sce.all.int,  features = gene_list, 
                group.by = 'RNA_snn_res.0.8') + 
    theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=0.5))
  p2    
  ggsave('pancreatic-0.8.pdf',width=12)
  
  
  pdf('orig.ident-vs-RNA_snn_res.0.1.pdf',height =  12 )
  gplots::balloonplot(table(sce.all.int$RNA_snn_res.0.1,sce.all.int$orig.ident))
  dev.off()
}


source('scRNA_scripts/lib.R')
sce.all.int = readRDS('2-harmony/sce.all_int.rds')
sp='human'
tmp=sce.all.int@meta.data
colnames(sce.all.int@meta.data) 

###查看默认Idents
Idents(sce.all.int)
table(sce.all.int$RNA_snn_res.0.5) 
###设置为RNA_snn_res.0.3继续后续的单细胞亚群注释
sel.clust = "RNA_snn_res.0.5"
sce.all.int <- SetIdent(sce.all.int, value = sel.clust)
table(sce.all.int@active.ident) 
table(sce.all.int$RNA_snn_res.0.5)
# pbmc_small <- BuildClusterTree(object = sce.all.int)
# plot(Tool(object = pbmc_small, slot = 'BuildClusterTree'))
# plot(pbmc_small@tools$BuildClusterTree)


# 付费环节 800 元人民币
# 如果是手动给各个单细胞亚群命名
###### 常见分群
# T Cells (CD3D, CD3E, CD8A), 
# B cells (CD19, CD79A, MS4A1 [CD20]), 
# Plasma cells (IGHG1, MZB1, SDC1, CD79A), 
# macrophages (CD68, CD163),
# 'CCL3L1' ,   #M2
# 'FABP4',  #M1
# Monocytes  (CD14),
# NK Cells (FGFBP2, FCG3RA, CX3CR1),  
# Fibroblasts (FGF7, MME), 
# Neutrophil ('G0S2', 'S100A9','S100A8','CXCL8')
# Endothelial cells (PECAM1, VWF). 
# epi or tumor (EPCAM, KRT19, PROM1, ALDH1A1, CD24).
# immune (CD45+,PTPRC), epithelial/cancer (EpCAM+,EPCAM), 
# stromal (CD10+,MME,fibo or CD31+,PECAM1,endo) 


#####注释亚群
#通常我们第一层次降维聚类分群：immune (CD45+,PTPRC),epithelial/cancer (EpCAM+,EPCAM),stromal (CD10+,MME,fibro or CD31+,PECAM1,endo)
#文章中一般第一次分群为常见的细胞群：上皮细胞（EPCAM、KRT19、CLDN4）、基质（PECAM1、CLO1A2、VWF）、增殖性（MKI67、STMN1、PCNA）、T（CD3D、CD3E、CD2）、B（CD79A，IGHG1，MS4A1），NK（KLRD1、GNLY、KLRF1）和髓系（CSF1R、CSF3R、CD68）细胞。

#绝大部分文章都是抓住免疫细胞亚群进行细分，包括淋巴系（T,B,NK细胞）和髓系（单核，树突，巨噬，粒细胞）的两大类作为第二次细分亚群。但是也有不少文章是抓住stromal 里面的 fibro 和endo进行细分
# genes_to_check <- c("PTPRC", "CD3D", "CD3E", "CD8A", "CD8B",'CD4', ##T
#                     "NKG7", "GNLY", "FCGR3", "ANCAM1", "CD160", #NK
#                     "CD19", "MS4A1", "CD79B", "CD79A", #B
#                     "MZB1", "XBP1", "JCHAIN", #Plasma
#                     "TPSAB1", "TPSB2",'CPA3','CD38','KIT','NR4A2','NR4A1','TNFSF10', #Mast
#                     "LYZ", "S100A8", "S100A9", "CD14", "CD68", #Myeloid
#                     "PPBP", "MPIG6B", #Platelets
#                     "HBB", "HBA1", "HBA2", #Erythrocytes
#                     "CXCL8", "DCN", "SLPI", "CXADR",'FAP','CFD','KRT8', #Fibroblasts
#                     "PECAM1", "VWF", "ENG", "MCAM", #Endothelial
#                     "MUC5AC", "TFF2", "TFF1", "GKN1", "GKN2", #Gastric
#                     "ALB", "APOC1", "APOA2", "ADH1B", "FGB", "FABP1", "ALDOB", #Hepatocytes
#                     "APOA1", "APOA4", "ALPI", "RBP2", "FABP2", "CDH17", "REG4", #Intestinal-like cells
#                     "EPCAM", "SOX9", "MET", "EGFR", "MSLN", "YAP1", #Malignant
#                     "MKI67", "BIRC5", "CDK1",'TOP2A') #Proliferative


genes_to_check <- c("KRT19", "EPCAM", "FXYD2", "ANXA4", "PRSS1", "CLPS", "PCSK1N", "TTR", "CTSS",
                              "FCN1", "CD68", "APOE", "CSF3R", "FCGR3B", "IRF8", "IRF7", "TPSAB1", 
                              "CPA3", "CD3D", "IL7R", "CD8A", "CCL5", "TIGIT", "FOXP3", "NKG7", 
                              "GZMA", "CD79A", "MS4A1", 'IGLL5','IGJ',
                              "STMN1", "MKI67", "PDGFRB", "TAGLN", "COL1A1", "COL1A2", "VWF", "CLDN5")
library(stringr)  
genes_to_check=str_to_upper(genes_to_check)
genes_to_check
rownames(sce.all.int)
length(rownames(sce.all.int))
#'KRT8' %in% rownames(sce.all.int)
p = DotPlot(sce.all.int, features = unique(genes_to_check),
            assay='RNA'  )  + coord_flip()
p


###重要：注意人工肉眼对应，需要根据不同数据集修改分群所对应的细胞亚群
#结合check-by-0.1文件夹中的marker基因和上述genes_to_check的dotplot
if(T){
  sce.all.int
  celltype=data.frame(ClusterID=0:23 ,
                      celltype=0:23 ) 
  #定义细胞亚群        
  celltype[celltype$ClusterID %in% c( 0,9,11 ),2]='Epithelial'
  celltype[celltype$ClusterID %in% c( 18  ),2]='Ductal'
  celltype[celltype$ClusterID %in% c(6),2]='Acinar'
  celltype[celltype$ClusterID %in% c(21 ),2]='Endocrine'
  celltype[celltype$ClusterID %in% c( 7 ),2]='Monocytes'
  celltype[celltype$ClusterID %in% c( 4,5 ),2]='Macrophages'
  celltype[celltype$ClusterID %in% c(1,17 ),2]='Neutrophils'
  ###可以看一下每个亚群的特异性基因表达top3
  #如果还不确定，可以走一下自动注释大致对应一下
  celltype[celltype$ClusterID %in% c( 22 ),2]='DC'  
  celltype[celltype$ClusterID %in% c( 8),2]='Mast'  
  celltype[celltype$ClusterID %in% c( 2,20 ),2]='CD4T'
  celltype[celltype$ClusterID %in% c( 19 ),2]='CD8T' 
  celltype[celltype$ClusterID %in% c(  ),2]='Tregs'  
  celltype[celltype$ClusterID %in% c( 3),2]='NK'  
  celltype[celltype$ClusterID %in% c( 15 ),2]='B'
  celltype[celltype$ClusterID %in% c(12,23 ),2]='Plasma' 
  celltype[celltype$ClusterID %in% c( 13),2]='Proliferative'  
  celltype[celltype$ClusterID %in% c( 10,14 ),2]='Fibroblasts'
  celltype[celltype$ClusterID %in% c( 16 ),2]='Endothelial' 
  
  head(celltype)
  celltype
  table(celltype$celltype)
  sce.all.int@meta.data$celltype = "NA"
  
  ###上面选择的是RNA_snn_res.0.5，如果是别的resolution，是需要修改的
  for(i in 1:nrow(celltype)){
    sce.all.int@meta.data[which(sce.all.int@meta.data$RNA_snn_res.0.5 == celltype$ClusterID[i]),'celltype'] <- celltype$celltype[i]}
  table(sce.all.int$celltype)
  Idents(sce.all.int)=sce.all.int$celltype
  
  
}

###查看亚群注释后marker表达是否对应的上，若不合理需重新检查上一步细胞亚群注释代码
p = DotPlot(sce.all.int, features = unique(genes_to_check),
            assay='RNA'  )  + coord_flip()
p

###去除未被注释上的细胞亚群
if("celltype" %in% colnames(sce.all.int@meta.data ) ){
  
  sel.clust = "celltype"
  sce.all.int <- SetIdent(sce.all.int, value = sel.clust)
  table(sce.all.int@active.ident) 
  
  phe=sce.all.int@meta.data
  save(phe,file = 'phe.Rdata')
  pdf('celltype-vs-orig.ident.pdf',width = 10)
  gplots::balloonplot(table(sce.all.int$celltype,sce.all.int$orig.ident))
  dev.off()
  dir.create('check-by-celltype')
  setwd('check-by-celltype')
  source('../scRNA_scripts/check-all-markers.R')
  setwd('../') 
  getwd()
}

library(paletteer)
color <- c(paletteer_d("awtools::bpalette"),
           paletteer_d("awtools::a_palette"),
           paletteer_d("awtools::mpalette"))
umap1=DimPlot(sce.all.int, reduction = "umap",
              # split.by = "orig.ident",
              group.by = "RNA_snn_res.0.5",label = T,cols = color)
umap1


umap2=DimPlot(sce.all.int, reduction = "umap",
              # split.by = "orig.ident",
              group.by = "celltype",label = T,repel = T,cols = color)

umap2
umap1+umap2
